xlint: add option to lint from git index
This commit introduces two new uses of xlint:

    $ xlint :PKGNAME

Lints the template corresponding to package PKGNAME as currently
staged in the git index (i.e. the version that would be committed)

    $ xlint :

Lints all templates that are changed in the git index.
Useful to run from the pre-commit hook for the void-packages repo.