xlint: allow comments on version definition lines
This is an oddly specific request that stems from my own CI setup. I use
renovatebot to update packages in a private fork of void-packages. To
simplify renovate's configuration, I write comments on the same line as
the version number. Unfortunately this can trigger an xlint error in
certain cases where it doesn't make sense, and this commit would fix
that.